ajustments

+++ include/processor.h
@@ -5,44 +5,44 @@
 
 class Processor {
  public:
-  float Utilization();  // TODO: See src/processor.cpp
+  double Utilization();  // TODO: See src/processor.cpp
   std::string getKey() const;
   int getPos() const;
-  std::ifstream& GotoLine(std::ifstream& file, unsigned int num);
-  void setUser(float u);
-  void setNice(float u);
-  void setSystem(float u);
-  void setIdle(float u);
-  void setIowait(float u);
-  void setIrq(float u);
-  void setSoftirq(float u);
-  void setSteal(float u);
+  std::ifstream& GotoLine(std::ifstream& file, int num);
+  void setUser(double u);
+  void setNice(double u);
+  void setSystem(double u);
+  void setIdle(double u);
+  void setIowait(double u);
+  void setIrq(double u);
+  void setSoftirq(double u);
+  void setSteal(double u);
 
-  float getUser();
-  float getNice();
-  float getSystem();
-  float getIdle();
-  float getIowait();
-  float getIrq();
-  float getSoftirq();
-  float getSteal();
+  double getUser();
+  double getNice();
+  double getSystem();
+  double getIdle();
+  double getIowait();
+  double getIrq();
+  double getSoftirq();
+  double getSteal();
 
 
-  Processor(int pos, std::string key) : pos_(pos), key_(key) {} 
- 
+  Processor(int pos, std::string key) : pos_(pos), key_(key) {}
+
 
   // TODO: Declare any necessary private members
  private:
     int pos_;
     std::string key_;
-    float user_ = {0.0};
-    float nice_ = {0.0};
-    float system_ = {0.0};
-    float idle_ = {0.0};
-    float iowait_ = {0.0};
-    float irq_ = {0.0};
-    float softirq_ = {0.0};
-    float steal_ = {0.0};
+    double user_ = {0.0};
+    double nice_ = {0.0};
+    double system_ = {0.0};
+    double idle_ = {0.0};
+    double iowait_ = {0.0};
+    double irq_ = {0.0};
+    double softirq_ = {0.0};
+    double steal_ = {0.0};
 };
 
 #endif
